requirements.txt
langchain>=0.1.0,<0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
openai>=1.0.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
qdrant-client~=1.9.0
streamlit>=1.20.0
beautifulsoup4>=4.0.0
lxml>=4.0.0
sqlalchemy>=2.0.0
sentence-transformers>=2.2.0
nest-asyncio>=1.5.0 
psycopg2-binary 
tavily-python
psycopg2-binary
sqlalchemy
langchain-experimental